+++ Minecoin/src/main/java/com/legopitstop/minecoin/item/DyeableCardItem.java
@@ -1,0 +1,9 @@
+package com.legopitstop.minecoin.item;
+
+import net.minecraft.item.DyeableItem;
+
+public class DyeableCardItem extends CardItem implements DyeableItem {
+    public DyeableCardItem(Settings settings) {
+        super(settings);
+    }
+}