+++ Minecoin/src/main/java/com/legopitstop/minecoin/util/EntityDataSaver.java
@@ -1,0 +1,7 @@
+package com.legopitstop.minecoin.util;
+
+import net.minecraft.nbt.NbtCompound;
+
+public interface EntityDataSaver {
+    NbtCompound getPersistentData();
+}