+++ More_Gold/src/main/java/com/legopitstop/moregold/MoreGoldDataGenerator.java
@@ -1,0 +1,24 @@
+package com.legopitstop.moregold;
+
+import com.legopitstop.moregold.data.MoreGoldWorldGenerator;
+import com.legopitstop.moregold.world.MoreGoldConfiguredFeatures;
+import com.legopitstop.moregold.world.MoreGoldPlacedFeatures;
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+import net.minecraft.registry.RegistryBuilder;
+import net.minecraft.registry.RegistryKeys;
+
+public class MoreGoldDataGenerator implements DataGeneratorEntrypoint {
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
+        FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
+
+        pack.addProvider(MoreGoldWorldGenerator::new);
+    }
+
+    @Override
+    public void buildRegistry(RegistryBuilder registryBuilder) {
+        registryBuilder.addRegistry(RegistryKeys.CONFIGURED_FEATURE, MoreGoldConfiguredFeatures::bootstrap);
+        registryBuilder.addRegistry(RegistryKeys.PLACED_FEATURE, MoreGoldPlacedFeatures::bootstrap);
+    }
+}