+++ More_Gold/src/main/java/com/legopitstop/moregold/world/tree/GoldenFruitSaplingGenerator.java
@@ -1,6 +1,6 @@
 package com.legopitstop.moregold.world.tree;
 
-import com.legopitstop.moregold.world.ModConfiguredFeatures;
+import com.legopitstop.moregold.world.MoreGoldConfiguredFeatures;
 import net.minecraft.block.sapling.SaplingGenerator;
 import net.minecraft.registry.RegistryKey;
 import net.minecraft.util.math.random.Random;
@@ -11,6 +11,6 @@ public class GoldenFruitSaplingGenerator extends SaplingGenerator {
     @Nullable
     @Override
     protected RegistryKey<ConfiguredFeature<?, ?>> getTreeFeature(Random random, boolean bees) {
-        return bees ? ModConfiguredFeatures.GOLDEN_FRUIT_BEES_005 : ModConfiguredFeatures.GOLDEN_FRUIT;
+        return bees ? MoreGoldConfiguredFeatures.GOLDEN_FRUIT_BEES_005 : MoreGoldConfiguredFeatures.GOLDEN_FRUIT;
     }
 }