+++ Spawner_Craft/src/main/java/com/legopitstop/spawnercraft/SpawnerCraftDataGenerator.java
@@ -1,0 +1,16 @@
+package com.legopitstop.spawnercraft;
+
+import net.fabricmc.fabric.api.datagen.v1.DataGeneratorEntrypoint;
+import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
+
+import com.legopitstop.spawnercraft.data.SpawnerCraftRecipeGenerator;
+
+public class SpawnerCraftDataGenerator implements DataGeneratorEntrypoint {
+    @Override
+    public void onInitializeDataGenerator(FabricDataGenerator fabricDataGenerator) {
+        FabricDataGenerator.Pack pack = fabricDataGenerator.createPack();
+
+        pack.addProvider(SpawnerCraftRecipeGenerator::new);
+
+    }
+}