+++ More_Gold/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - dd/mm/yyyy
+## [1.0.1] - unreleased
+### General
+- Saplings, and Seeds can now be composted.
+- Golden Logs & Wood can now be used as fuel.
+
+### Bug Fixes
+- #2 Black on Golden Fruit Crop & Golden Fruit Sapling
+
+## [1.0.0] - 7/31/2023
 
 Initial release